+++ app/services/storage/models/blobStorageResult.ts
